+++ packages/core/dynamic-table/src/components/TableHead.tsx
@@ -13,6 +13,7 @@ interface Props {
   onSort: (item: RowCellType) => () => void;
   isRankable?: boolean;
   isRanking: boolean;
+  testId?: string;
 }
 
 class TableHead extends React.Component<Props, {}> {
@@ -41,6 +42,7 @@ class TableHead extends React.Component<Props, {}> {
       onSort,
       isRanking,
       isRankable,
+      testId,
     } = this.props;
 
     if (!head) {
@@ -52,7 +54,11 @@ class TableHead extends React.Component<Props, {}> {
     const { cells, ...rest } = head;
 
     return (
-      <Head {...rest} isRanking={isRanking}>
+      <Head
+        {...rest}
+        isRanking={isRanking}
+        data-testid={testId && `${testId}--head`}
+      >
         <tr>
           {cells.map((cell, index) => {
             const { isSortable, key, ...restCellProps } = cell;
@@ -69,6 +75,7 @@ class TableHead extends React.Component<Props, {}> {
                     ? onSort(cell)()
                     : undefined
                 }
+                testId={testId}
                 sortOrder={key === sortKey ? sortOrder : undefined}
                 {...restCellProps}
               />