+++ packages/core/dynamic-table/src/components/rankable/TableRow.tsx
@@ -13,6 +13,7 @@ export interface Props extends WithDimensionsProps {
   rowIndex: number;
   isRankingDisabled: boolean;
   isHighlighted?: boolean;
+  testId?: string;
 }
 
 export class RankableTableRow extends React.Component<Props, {}> {
@@ -33,6 +34,7 @@ export class RankableTableRow extends React.Component<Props, {}> {
       rowIndex,
       isRankingDisabled,
       isHighlighted,
+      testId,
     } = this.props;
     const { cells, key, ...restRowProps } = row;
     const inlineStyles = inlineStylesIfRanking(isRanking, refWidth);
@@ -69,6 +71,7 @@ export class RankableTableRow extends React.Component<Props, {}> {
                   isRanking={isRanking}
                   key={cellIndex} // eslint-disable-line react/no-array-index-key
                   isFixedSize={isFixedSize}
+                  testId={testId}
                 />
               );
             })}