+++ packages/editor/adf-schema/src/__tests__/unit/json-schema/backwards-compat.ts
@@ -24,15 +24,9 @@ function fetchLastPublishedJSONSchema() {
   );
 }
 
-// function fetchMasterJSONSchema() {
-//   return fetchJSONSchema(
-//     `https://bitbucket.org/atlassian/atlaskit-mk-2/raw/master/packages/editor/adf-schema/json-schema/v1/full.json`,
-//   );
-// }
-
-function fetchEditorNextReleaseJSONSchema() {
+function fetchMasterJSONSchema() {
   return fetchJSONSchema(
-    `https://bitbucket.org/atlassian/atlaskit-mk-2/raw/editor-next-release/packages/editor/adf-schema/json-schema/v1/full.json`,
+    `https://bitbucket.org/atlassian/atlaskit-mk-2/raw/master/packages/editor/adf-schema/json-schema/v1/full.json`,
   );
 }
 
@@ -65,7 +59,7 @@ describe('JSON schema', () => {
       // This should prevent breakcing changes are introduced into editor-next-release
       // We should revert this back to fetchMasterJSONSchema when editor-next-release is merged back
       // into master
-      existingSchema = await fetchEditorNextReleaseJSONSchema();
+      existingSchema = await fetchMasterJSONSchema();
     } catch (err) {
       // if package with this version doesn't exist test against the latest version
       // this can happen when you manually bump version in package.json