+++ packages/editor/editor-core/examples/5-full-page.tsx
@@ -228,6 +228,7 @@ export class ExampleEditorComponent extends React.Component<
                 component: ExampleInlineCommentComponent,
               }}
               allowStatus={true}
+              allowNestedTasks
               {...providers}
               media={{
                 provider: mediaProvider,