+++ packages/editor/editor-core/src/plugins/table/pm-plugins/table-resizing/event-handlers.ts
@@ -19,13 +19,13 @@ import {
 import { getSelectedColumnIndexes, updateResizeHandles } from '../../utils';
 import { pluginKey as widthPluginKey } from '../../../width';
 import { getPluginState } from './plugin';
-import { setDragging, evenColumns } from './commands';
+import { setDragging, evenColumns, stopResizing } from './commands';
 import { getParentNodeWidth } from '../../../../utils/node-width';
 
 export const handleMouseDown = (
   view: EditorView,
   event: MouseEvent,
-  resizeHandlePos: number,
+  localResizeHandlePos: number,
   dynamicTextSizing: boolean,
 ): boolean => {
   const { state, dispatch } = view;
@@ -34,14 +34,14 @@ export const handleMouseDown = (
 
   if (
     editorDisabled ||
-    resizeHandlePos === null ||
-    !pointsAtCell(state.doc.resolve(resizeHandlePos))
+    localResizeHandlePos === null ||
+    !pointsAtCell(state.doc.resolve(localResizeHandlePos))
   ) {
     return false;
   }
   event.preventDefault();
-  const cell = state.doc.nodeAt(resizeHandlePos);
-  const $cell = state.doc.resolve(resizeHandlePos);
+  const cell = state.doc.nodeAt(localResizeHandlePos);
+  const $cell = state.doc.resolve(localResizeHandlePos);
   const originalTable = $cell.node(-1);
   const start = $cell.start(-1);
 
@@ -88,7 +88,7 @@ export const handleMouseDown = (
     return true;
   }
 
-  const width = currentColWidth(view, resizeHandlePos, cell!
+  const width = currentColWidth(view, localResizeHandlePos, cell!
     .attrs as CellAttributes);
 
   setDragging({ startX: event.clientX, startWidth: width })(state, dispatch);
@@ -99,11 +99,12 @@ export const handleMouseDown = (
 
     const { clientX } = event;
     const { state, dispatch } = view;
-    const { dragging } = getPluginState(state);
-    if (
-      resizeHandlePos === null ||
-      !pointsAtCell(state.doc.resolve(resizeHandlePos))
-    ) {
+    const { dragging, resizeHandlePos } = getPluginState(state);
+    if (resizeHandlePos === null) {
+      return stopResizing()(state, dispatch);
+    }
+
+    if (!pointsAtCell(state.doc.resolve(resizeHandlePos))) {
       return;
     }
     // resizeHandlePos could be remapped via a collab change.
@@ -114,8 +115,7 @@ export const handleMouseDown = (
 
     // If we let go in the same place we started, dont need to do anything.
     if (dragging && clientX === dragging.startX) {
-      setDragging(null)(state, dispatch);
-      return;
+      return stopResizing()(state, dispatch);
     }
 
     let { tr } = state;
@@ -149,14 +149,14 @@ export const handleMouseDown = (
         tr = updateColumnWidths(newResizeState, table, start)(tr);
       }
 
-      setDragging(null, tr)(view.state, dispatch);
+      return stopResizing(tr)(state, dispatch);
     }
   }
 
   function move(event: MouseEvent) {
     const { clientX, which } = event;
     const { state } = view;
-    const { dragging } = getPluginState(state);
+    const { dragging, resizeHandlePos } = getPluginState(state);
     if (
       !which ||
       !dragging ||