+++ packages/elements/mention/src/__tests__/unit/api/MentionResourceSpec.ts
@@ -29,8 +29,22 @@ const options = (
   omitCredentials,
 });
 
-const getSecurityHeader = (call: any) =>
-  call[1].headers.get(defaultSecurityHeader);
+type MockFetchResponse = Array<
+  | string
+  | {
+      headers: any;
+      credentials: string;
+      request: any;
+      identifier: string;
+      isUnmatched: any;
+    }
+>;
+
+const getSecurityHeader = (call: MockFetchResponse) => {
+  const data = call[1];
+  if (typeof data === 'string') return undefined;
+  return data.headers[defaultSecurityHeader];
+};
 
 const defaultSecurityCode = '10804';
 
@@ -380,7 +394,7 @@ describe('MentionResource', () => {
             const calls = fetchMock.calls(matcher.name);
             expect(calls).toHaveLength(2);
             expect(getSecurityHeader(calls[0])).toEqual(defaultSecurityCode);
-            expect(getSecurityHeader(calls[1])).toEqual('666');
+            expect(getSecurityHeader(calls[1])).toEqual(666);
             done();
           } catch (ex) {
             done.fail(ex);