+++ packages/elements/task-decision/src/components/ResourcedTaskItem.tsx
@@ -7,7 +7,6 @@ import {
   ContentRef,
   TaskDecisionProvider,
   TaskState,
-  User,
   DecisionState,
 } from '../types';
 import { FabricElementsAnalyticsContext } from '@atlaskit/analytics-namespaced-context';
@@ -20,18 +19,14 @@ export interface Props {
   children?: any;
   taskDecisionProvider?: Promise<TaskDecisionProvider>;
   objectAri?: string;
-  containerAri?: string;
   showPlaceholder?: boolean;
   placeholder?: string;
   appearance?: Appearance;
-  creator?: User;
-  lastUpdater?: User;
   disabled?: boolean;
 }
 
 export interface State {
   isDone?: boolean;
-  lastUpdater?: User;
 }
 
 export default class ResourcedTaskItem extends PureComponent<Props, State> {
@@ -45,7 +40,6 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
 
     this.state = {
       isDone: props.isDone,
-      lastUpdater: props.lastUpdater,
     };
   }
 
@@ -53,7 +47,6 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
     this.mounted = true;
     this.subscribe(
       this.props.taskDecisionProvider,
-      this.props.containerAri,
       this.props.objectAri,
       this.props.isDone,
     );
@@ -68,13 +61,11 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
     }
     if (
       nextProps.taskDecisionProvider !== this.props.taskDecisionProvider ||
-      nextProps.containerAri !== this.props.containerAri ||
       nextProps.objectAri !== this.props.objectAri
     ) {
       this.unsubscribe();
       this.subscribe(
         nextProps.taskDecisionProvider,
-        nextProps.containerAri,
         nextProps.objectAri,
         nextProps.isDone,
       );
@@ -88,45 +79,32 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
 
   private subscribe(
     taskDecisionProvider?: Promise<TaskDecisionProvider>,
-    containerAri?: string,
     objectAri?: string,
     isDone?: boolean,
   ) {
-    if (taskDecisionProvider && containerAri && objectAri) {
+    if (taskDecisionProvider && objectAri) {
       taskDecisionProvider.then(provider => {
         if (!this.mounted) {
           return;
         }
         const { taskId } = this.props;
-        const objectKey = { localId: taskId, objectAri, containerAri };
+        const objectKey = { localId: taskId, objectAri };
         const item: BaseItem<TaskState> = {
           ...objectKey,
           state: isDone ? 'DONE' : 'TODO',
           lastUpdateDate: new Date(),
           type: 'TASK',
         };
-        provider.subscribe(
-          { localId: taskId, objectAri, containerAri },
-          this.onUpdate,
-          item,
-        );
+        provider.subscribe({ localId: taskId, objectAri }, this.onUpdate, item);
       });
     }
   }
 
   private unsubscribe() {
-    const {
-      taskDecisionProvider,
-      taskId,
-      objectAri,
-      containerAri,
-    } = this.props;
-    if (taskDecisionProvider && containerAri && objectAri) {
+    const { taskDecisionProvider, taskId, objectAri } = this.props;
+    if (taskDecisionProvider && objectAri) {
       taskDecisionProvider.then(provider => {
-        provider.unsubscribe(
-          { localId: taskId, objectAri, containerAri },
-          this.onUpdate,
-        );
+        provider.unsubscribe({ localId: taskId, objectAri }, this.onUpdate);
       });
     }
   }
@@ -136,23 +114,18 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
   };
 
   private handleOnChange = (taskId: string, isDone: boolean) => {
-    const {
-      taskDecisionProvider,
-      objectAri,
-      containerAri,
-      onChange,
-    } = this.props;
+    const { taskDecisionProvider, objectAri, onChange } = this.props;
     // Optimistically update the task
     this.setState({ isDone });
 
-    if (taskDecisionProvider && containerAri && objectAri) {
+    if (taskDecisionProvider && objectAri) {
       // Call provider to update task
       taskDecisionProvider.then(provider => {
         if (!this.mounted) {
           return;
         }
         provider.toggleTask(
-          { localId: taskId, objectAri, containerAri },
+          { localId: taskId, objectAri },
           isDone ? 'DONE' : 'TODO',
         );
 
@@ -160,16 +133,6 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
         // we should only call onChange once the internal state have been modified
         if (onChange) {
           onChange(taskId, isDone);
-        }
-
-        if (isDone) {
-          // Undefined provider.getCurrentUser or currentUser shows 'Created By'
-          // ie. does not update to prevent incorrect 'Completed By' message
-          this.setState({
-            lastUpdater: provider.getCurrentUser
-              ? provider.getCurrentUser()
-              : undefined,
-          });
         }
       });
     } else {
@@ -181,13 +144,11 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
   };
 
   render() {
-    const { isDone, lastUpdater } = this.state;
+    const { isDone } = this.state;
     const {
       appearance,
       children,
-      containerAri,
       contentRef,
-      creator,
       objectAri,
       showPlaceholder,
       placeholder,
@@ -198,7 +159,6 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
     return (
       <FabricElementsAnalyticsContext
         data={{
-          containerAri,
           objectAri,
         }}
       >
@@ -210,8 +170,6 @@ export default class ResourcedTaskItem extends PureComponent<Props, State> {
           contentRef={contentRef}
           showPlaceholder={showPlaceholder}
           placeholder={placeholder}
-          creator={creator}
-          lastUpdater={lastUpdater}
           disabled={disabled}
         >
           {children}