+++ packages/elements/task-decision/src/type-helpers.ts
@@ -3,11 +3,8 @@ import {
   Decision,
   Item,
   ObjectKey,
-  RendererContext,
   ServiceDecision,
-  ServiceItem,
   ServiceTask,
-  SortCriteria,
   Task,
   TaskState,
 } from './types';
@@ -17,38 +14,19 @@ export const isDecision = (item: Item): item is Decision =>
 export const isTask = (item: Item): item is Task =>
   !!(item && item.type === 'TASK');
 
-export const isServiceDecision = (item: ServiceItem): item is ServiceDecision =>
-  !!(item && item.type === 'DECISION');
-export const isServiceTask = (item: ServiceItem): item is ServiceTask =>
-  !!(item && item.type === 'TASK');
-
-export const isDateSortCriteria = (sortCriteria?: SortCriteria) =>
-  !sortCriteria ||
-  sortCriteria === 'creationDate' ||
-  sortCriteria === 'lastUpdateDate';
-
 export const toObjectKey = (
   item: Item | ServiceDecision | ServiceTask | BaseItem<any>,
 ): ObjectKey => {
-  const { containerAri, localId, objectAri } = item;
+  const { localId, objectAri } = item;
   return {
-    containerAri,
     localId,
     objectAri,
   };
 };
 
-export const toRendererContext = (item: Item | ObjectKey): RendererContext => {
-  const { containerAri, objectAri } = item;
-  return {
-    containerAri,
-    objectAri,
-  };
-};
-
 export const objectKeyToString = (objectKey: ObjectKey) => {
-  const { containerAri, objectAri, localId } = objectKey;
-  return `${containerAri}:${objectAri}:${localId}`;
+  const { objectAri, localId } = objectKey;
+  return `${objectAri}:${localId}`;
 };
 
 export const toggleTaskState = (state: TaskState): TaskState =>