+++ packages/media/media-card/src/files/cardImageView/index.tsx
@@ -36,6 +36,7 @@ export interface FileCardImageViewProps {
   readonly fileSize?: string;
 
   readonly dataURI?: string;
+  readonly alt?: string;
   readonly progress?: number;
   readonly status: CardStatus;
 
@@ -50,6 +51,7 @@ export interface FileCardImageViewProps {
 
   readonly actions?: CardAction[];
   readonly onRetry?: () => void;
+  readonly onDisplayImage?: () => void;
   readonly previewOrientation?: number;
 }
 
@@ -57,6 +59,7 @@ export class FileCardImageViewBase extends Component<
   FileCardImageViewProps & WithAnalyticsEventsProps,
   {}
 > {
+  private wasThumbnailDisplayed = false;
   private lastAnalyticsAction: AnalyticsLoadingAction | undefined;
   static defaultProps = {
     resizeMode: 'crop',
@@ -182,16 +185,32 @@ export class FileCardImageViewBase extends Component<
   };
 
   private renderMediaImage = () => {
-    const { dataURI, mediaType, previewOrientation } = this.props;
+    const {
+      dataURI,
+      mediaType,
+      previewOrientation,
+      onDisplayImage,
+      alt,
+    } = this.props;
 
     if (!shouldDisplayImageThumbnail(dataURI, mediaType)) {
       this.fireLoadingStatusAnalyticsEvent('succeeded');
       return null;
     }
 
+    if (
+      !this.wasThumbnailDisplayed &&
+      onDisplayImage &&
+      mediaType === 'image'
+    ) {
+      onDisplayImage();
+      this.wasThumbnailDisplayed = true;
+    }
+
     return (
       <MediaImage
         dataURI={dataURI}
+        alt={alt}
         crop={this.isCropped}
         stretch={this.isStretched}
         previewOrientation={previewOrientation}